refactor: api, web 및 util의 패키지 구조 변경
