+++ src/main/java/org/weekly/weekly/VoucherManagementController.java
@@ -9,10 +9,10 @@ import org.weekly.weekly.customer.dto.request.CustomerUpdateRequest;
 import org.weekly.weekly.customer.dto.response.CustomerResponse;
 import org.weekly.weekly.customer.dto.response.CustomersResponse;
 import org.weekly.weekly.ui.CommandLineApplication;
-import org.weekly.weekly.util.CustomerMenu;
-import org.weekly.weekly.util.ManageMenu;
-import org.weekly.weekly.util.PrintMessageType;
-import org.weekly.weekly.util.VoucherMenu;
+import org.weekly.weekly.global.util.CustomerMenu;
+import org.weekly.weekly.global.util.ManageMenu;
+import org.weekly.weekly.global.util.PrintMessageType;
+import org.weekly.weekly.global.util.VoucherMenu;
 import org.weekly.weekly.voucher.controller.VoucherController;
 import org.weekly.weekly.voucher.dto.request.VoucherCreationRequest;
 import org.weekly.weekly.voucher.dto.response.VoucherCreationResponse;
