+++ src/main/java/org/weekly/weekly/customer/exception/CustomerException.java
@@ -1,6 +1,6 @@
 package org.weekly.weekly.customer.exception;
 
-import org.weekly.weekly.util.ExceptionCode;
+import org.weekly.weekly.global.handler.ExceptionCode;
 
 public class CustomerException extends RuntimeException {
