+++ src/main/java/org/weekly/weekly/global/util/ManageMenu.java
@@ -1,4 +1,6 @@
-package org.weekly.weekly.util;
+package org.weekly.weekly.global.util;
+
+import org.weekly.weekly.global.handler.ExceptionCode;
 
 import java.util.Arrays;
 import java.util.Map;
