+++ src/main/java/org/weekly/weekly/global/util/Menu.java
@@ -1,4 +1,4 @@
-package org.weekly.weekly.util;
+package org.weekly.weekly.global.util;
 
 public interface Menu {
     String printMessage();
