+++ src/main/java/org/weekly/weekly/ui/CommandLineApplication.java
@@ -9,10 +9,10 @@ import org.weekly.weekly.customer.dto.response.CustomersResponse;
 import org.weekly.weekly.ui.exception.InputValidator;
 import org.weekly.weekly.ui.reader.CommandReader;
 import org.weekly.weekly.ui.writer.SystemWriter;
-import org.weekly.weekly.util.CustomerMenu;
-import org.weekly.weekly.util.ManageMenu;
-import org.weekly.weekly.util.Menu;
-import org.weekly.weekly.util.VoucherMenu;
+import org.weekly.weekly.global.util.CustomerMenu;
+import org.weekly.weekly.global.util.ManageMenu;
+import org.weekly.weekly.global.util.Menu;
+import org.weekly.weekly.global.util.VoucherMenu;
 import org.weekly.weekly.voucher.domain.DiscountType;
 import org.weekly.weekly.voucher.dto.request.VoucherCreationRequest;
 import org.weekly.weekly.voucher.dto.request.VoucherInfoRequest;
