+++ src/main/java/org/weekly/weekly/ui/exception/InputException.java
@@ -1,6 +1,6 @@
 package org.weekly.weekly.ui.exception;
 
-import org.weekly.weekly.util.ExceptionCode;
+import org.weekly.weekly.global.handler.ExceptionCode;
 
 public class InputException extends RuntimeException {
