+++ src/main/java/org/weekly/weekly/ui/reader/BufferedReaderWrap.java
@@ -3,7 +3,7 @@ package org.weekly.weekly.ui.reader;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Component;
 import org.weekly.weekly.ui.exception.InputException;
-import org.weekly.weekly.util.ExceptionCode;
+import org.weekly.weekly.global.handler.ExceptionCode;
 
 import java.io.BufferedReader;
 import java.io.IOException;
