+++ src/main/java/org/weekly/weekly/voucher/repository/MemoryVoucherRepository.java
@@ -2,7 +2,7 @@ package org.weekly.weekly.voucher.repository;
 
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Repository;
-import org.weekly.weekly.util.ExceptionCode;
+import org.weekly.weekly.global.handler.ExceptionCode;
 import org.weekly.weekly.voucher.domain.DiscountType;
 import org.weekly.weekly.voucher.domain.Voucher;
 import org.weekly.weekly.voucher.exception.VoucherException;
