+++ src/test/java/org/weekly/weekly/util/VoucherMenuTest.java
@@ -2,6 +2,7 @@ package org.weekly.weekly.util;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import org.weekly.weekly.global.util.VoucherMenu;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
